requirements.txt
aiogram==3.15.0
python-dotenv==1.2.1
yt-dlp
Pillow
mutagen
requests